fix: cast bias to weight.dtype in fused_obs_encode fallback
F.linear fails when input and bias have different dtypes in PyTorch 2.x.
Cast bias to match weight dtype in the fallback path.

+++ pufferlib_market/kernels/fused_obs_encode.py
@@ -252,7 +252,7 @@ def _fused_obs_norm_linear_relu_fallback(
 ) -> torch.Tensor:
     """Pure PyTorch fallback: normalize then linear+relu."""
     obs_norm = (obs.float() - mean.float()) / (std.float() + eps)
-    h = F.linear(obs_norm.to(weight.dtype), weight, bias)
+    h = F.linear(obs_norm.to(weight.dtype), weight, bias.to(weight.dtype))
     return F.relu(h)
